app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-secret-manager==2.22.0
google-cloud-aiplatform==1.71.1
google-genai==1.29.0
httpx==0.27.2
